Responses to PR comments.
Refactored StorageUtil to a different class, changed pre-existing
behavior of silent failure on invalid bucket uri.
