+++ src/main/resources/com/google/jenkins/plugins/storage/Messages.properties
@@ -28,11 +28,11 @@ ExpiringBucketLifecycleManager.DetailsMessage=Contents expire in {0} days
 AbstractBucketLifecycleManagerDescriptor.MultiPartBucket=Bucket name cannot contain multiple parts: {0}
 AbstractUpload.IncludeException=Exception while resolving inclusions.
 AbstractUpload.UploadException=Exception while performing upload
-GoogleCloudStorageUploader.DisplayName=Google Cloud Storage Uploader
+GoogleCloudStorageUploader.DisplayName=Google Cloud Storage Plugin
 ClassicUpload.NoArtifacts=No artifacts found matching pattern: {0}
 GoogleCloudStorageUploader.ExceptionDuringUpload=Exception while uploading files: {0}
 ClassicUpload.FoundForPattern=Found {0} files to upload from pattern: {1}
-UploadModule.PrefixFormat=[{0}] {1}
+StorageUtil.PrefixFormat=[{0}] {1}
 AbstractUpload.Uploading=Uploading: {0}
 AbstractUpload.ExceptionFileUpload=exception uploading file
 AbstractUpload.RemoteCredentialError=An error occured preparing to use the provided credential remotely
